feat: add total blocks minted state on SPO state which relies on epoch activity counter
- Introduced a new query  in  to retrieve total blocks minted for specified pools.
- Updated the  and  to handle the new query and integrate it with the state management.
- Adjusted REST handlers to return the updated blocks minted data for pools.

+++ modules/rest_blockfrost/src/types.rs
@@ -15,7 +15,7 @@ pub struct PoolExtendedRest {
     pub hex: String,
     pub active_stake: String, // u64 in string
     pub live_stake: String,   // u64 in string
-    pub blocks_minted: usize,
+    pub blocks_minted: u64,
     pub live_saturation: Decimal,
     pub declared_pledge: String, // u64 in string
     pub margin_cost: f32,
